Camera draw arc/circle + improvements

+++ utility/number/number.go
@@ -248,7 +248,7 @@ func ValueMinimum[T Number]() T {
 	case uint64:
 		return any(0).(T)
 	case float32:
-		return any(-math.MaxFloat32).(T)
+		return any(float32(-math.MaxFloat32)).(T)
 	case float64:
 		return any(-math.MaxFloat64).(T)
 	default:
